Reorganized project structure and renamed files with proper naming protocol
