+++ cardiacengineering_app/lib/pages/home_data/gpm_page.dart
@@ -1,6 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'loginPage.dart';
 
 class GPMpage extends StatelessWidget {
   const GPMpage({Key? key}) : super(key: key);
