--- cardiacengineering_app/python/csv/data_1.csv
@@ -1,181 +1,0 @@
-Second,RPM,Battery,Flow
-0,79,100,100
-1,79,99,99
-2,76,99,99
-3,73,99,100
-4,70,99,101
-5,67,99,102
-6,67,99,101
-7,64,99,100
-8,61,99,99
-9,61,99,98
-10,61,99,99
-11,61,99,98
-12,61,99,97
-13,61,99,97
-14,61,99,98
-15,61,99,97
-16,61,99,96
-17,61,99,96
-18,61,99,97
-19,64,99,97
-20,64,99,98
-21,61,99,97
-22,61,99,96
-23,64,99,95
-24,61,99,95
-25,64,99,95
-26,64,99,96
-27,67,99,96
-28,67,99,96
-29,67,99,96
-30,67,99,95
-31,64,99,96
-32,64,99,95
-33,67,99,95
-34,64,99,95
-35,61,99,96
-36,64,99,97
-37,67,98,98
-38,64,98,98
-39,64,98,97
-40,67,98,96
-41,67,98,95
-42,64,98,95
-43,61,98,95
-44,64,98,95
-45,61,98,95
-46,61,98,95
-47,61,98,96
-48,61,98,96
-49,61,98,97
-50,64,98,97
-51,61,98,97
-52,64,98,96
-53,61,98,95
-54,61,98,96
-55,64,98,96
-56,64,98,97
-57,67,98,97
-58,64,98,96
-59,67,98,95
-60,70,98,96
-61,67,98,97
-62,70,98,97
-63,73,98,97
-64,76,98,97
-65,73,98,96
-66,70,98,97
-67,73,98,96
-68,73,98,96
-69,70,98,95
-70,73,98,96
-71,73,98,95
-72,76,98,95
-73,76,97,96
-74,79,97,96
-75,82,97,95
-76,82,97,95
-77,79,97,95
-78,79,97,95
-79,76,97,95
-80,76,97,96
-81,76,97,95
-82,79,97,96
-83,82,97,97
-84,85,97,97
-85,88,97,97
-86,85,97,97
-87,85,97,97
-88,88,97,96
-89,88,97,96
-90,85,97,96
-91,82,97,96
-92,79,97,95
-93,76,97,95
-94,73,97,95
-95,73,97,95
-96,73,97,96
-97,70,97,97
-98,67,97,97
-99,64,97,96
-100,64,97,95
-101,61,97,96
-102,64,97,96
-103,64,97,97
-104,61,97,98
-105,64,97,99
-106,61,97,98
-107,61,97,98
-108,64,97,98
-109,67,96,98
-110,67,96,97
-111,67,96,98
-112,70,96,99
-113,67,96,100
-114,64,96,99
-115,67,96,99
-116,64,96,100
-117,64,96,99
-118,67,96,100
-119,67,96,100
-120,64,96,100
-121,67,96,99
-122,67,96,100
-123,70,96,99
-124,67,96,100
-125,70,96,99
-126,73,96,99
-127,73,96,99
-128,73,96,99
-129,70,96,99
-130,73,96,100
-131,73,96,100
-132,73,96,101
-133,70,96,100
-134,67,96,100
-135,64,96,99
-136,64,96,100
-137,61,96,100
-138,64,96,99
-139,67,96,100
-140,67,96,101
-141,64,96,101
-142,64,96,100
-143,61,96,101
-144,61,96,101
-145,64,95,100
-146,64,95,100
-147,67,95,99
-148,67,95,98
-149,64,95,97
-150,61,95,96
-151,64,95,95
-152,61,95,95
-153,64,95,95
-154,64,95,95
-155,61,95,95
-156,61,95,95
-157,64,95,95
-158,67,95,96
-159,70,95,97
-160,70,95,96
-161,70,95,96
-162,67,95,97
-163,64,95,97
-164,61,95,96
-165,64,95,97
-166,64,95,98
-167,67,95,99
-168,70,95,100
-169,67,95,101
-170,70,95,100
-171,73,95,99
-172,76,95,100
-173,76,95,100
-174,73,95,101
-175,73,95,100
-176,73,95,100
-177,76,95,101
-178,73,95,101
-179,73,95,102
