--- cardiacengineering_app/python/csv/data_2.csv
@@ -1,181 +1,0 @@
-Second,RPM,Battery,Flow
-0,69,98,100
-1,66,97,99
-2,69,97,100
-3,66,97,99
-4,66,97,98
-5,69,97,99
-6,72,97,99
-7,75,97,100
-8,75,97,101
-9,75,97,100
-10,75,97,100
-11,75,97,100
-12,78,97,99
-13,81,97,98
-14,81,97,98
-15,81,97,98
-16,81,97,99
-17,78,97,100
-18,81,97,101
-19,81,97,101
-20,81,97,100
-21,78,97,99
-22,81,97,99
-23,84,97,99
-24,84,97,99
-25,87,97,98
-26,84,97,97
-27,84,97,96
-28,84,97,96
-29,87,97,95
-30,90,97,96
-31,87,97,96
-32,84,97,95
-33,87,97,95
-34,84,97,96
-35,87,97,96
-36,87,97,96
-37,84,96,95
-38,87,96,95
-39,90,96,95
-40,87,96,96
-41,87,96,95
-42,84,96,96
-43,84,96,97
-44,84,96,96
-45,87,96,97
-46,90,96,98
-47,93,96,98
-48,90,96,99
-49,93,96,100
-50,90,96,101
-51,87,96,102
-52,90,96,101
-53,87,96,100
-54,90,96,100
-55,87,96,99
-56,84,96,99
-57,87,96,99
-58,90,96,99
-59,87,96,99
-60,84,96,100
-61,84,96,101
-62,87,96,102
-63,84,96,102
-64,81,96,101
-65,84,96,101
-66,87,96,100
-67,84,96,99
-68,84,96,99
-69,87,96,99
-70,84,96,100
-71,87,96,101
-72,90,96,102
-73,93,95,103
-74,96,95,103
-75,93,95,104
-76,96,95,103
-77,96,95,102
-78,93,95,101
-79,90,95,100
-80,90,95,99
-81,90,95,98
-82,93,95,99
-83,96,95,98
-84,93,95,97
-85,90,95,98
-86,93,95,99
-87,90,95,98
-88,90,95,98
-89,90,95,99
-90,90,95,100
-91,90,95,101
-92,90,95,100
-93,87,95,101
-94,84,95,101
-95,81,95,100
-96,84,95,101
-97,84,95,101
-98,84,95,102
-99,87,95,102
-100,87,95,103
-101,90,95,102
-102,90,95,101
-103,93,95,101
-104,93,95,100
-105,93,95,101
-106,96,95,102
-107,96,95,102
-108,93,95,101
-109,93,94,101
-110,90,94,100
-111,93,94,101
-112,96,94,102
-113,93,94,102
-114,90,94,101
-115,90,94,101
-116,93,94,101
-117,96,94,102
-118,99,94,101
-119,99,94,101
-120,99,94,100
-121,99,94,101
-122,99,94,102
-123,96,94,103
-124,99,94,102
-125,96,94,101
-126,99,94,100
-127,99,94,100
-128,96,94,99
-129,99,94,99
-130,99,94,100
-131,96,94,101
-132,96,94,101
-133,93,94,101
-134,96,94,101
-135,93,94,102
-136,93,94,101
-137,90,94,102
-138,87,94,102
-139,84,94,102
-140,84,94,102
-141,81,94,103
-142,78,94,102
-143,75,94,103
-144,75,94,102
-145,72,93,102
-146,69,93,101
-147,66,93,101
-148,66,93,102
-149,69,93,102
-150,72,93,103
-151,69,93,104
-152,66,93,103
-153,63,93,102
-154,60,93,101
-155,60,93,100
-156,60,93,101
-157,60,93,100
-158,60,93,101
-159,60,93,102
-160,63,93,102
-161,63,93,102
-162,60,93,102
-163,60,93,102
-164,60,93,103
-165,60,93,103
-166,60,93,102
-167,60,93,101
-168,63,93,101
-169,60,93,101
-170,60,93,100
-171,60,93,100
-172,63,93,99
-173,66,93,99
-174,63,93,100
-175,63,93,101
-176,60,93,101
-177,60,93,101
-178,63,93,102
-179,66,93,101
